2회차
ㅣㄱ모찌

+++ Assets/Script/UnitBooster.cs
@@ -3,34 +3,33 @@ using UnityEngine;
 public class UnitBooster : MonoBehaviour
 {
     public Rigidbody2D rb;
-    public Transform B;
+    public Unit B;
     public int Qx; // 제 x사분면
     public float Rt;
     public float thrustPower = 1f;
     public float thrustSpeed = 0;
-    
     List<int> Q;
-    void Start ()
+    void Start()
     {
         rb = transform.parent.GetComponent<Rigidbody2D>();
-        B = transform.parent;
-
-        Q = new(){1,0,0,1};
-        if(transform.localPosition.x != 0 && transform.localPosition.y !=0)
+        B = transform.parent.GetComponent<Unit>();
+        
+        Q = new() { 1, 0, 0, 1 };
+        if (transform.localPosition.x != 0 && transform.localPosition.y != 0)
         {
-            if(transform.localPosition.x > 0 && transform.localPosition.y > 0)
+            if (transform.localPosition.x > 0 && transform.localPosition.y > 0)
             {
                 Qx = 1;
             }
-            else if(transform.localPosition.x < 0 && transform.localPosition.y > 0)
+            else if (transform.localPosition.x < 0 && transform.localPosition.y > 0)
             {
                 Qx = 2;
             }
-            else if(transform.localPosition.x < 0 && transform.localPosition.y < 0)
+            else if (transform.localPosition.x < 0 && transform.localPosition.y < 0)
             {
                 Qx = 3;
             }
-            else if(transform.localPosition.x > 0 && transform.localPosition.y < 0)
+            else if (transform.localPosition.x > 0 && transform.localPosition.y < 0)
             {
                 Qx = 4;
             }
@@ -39,35 +38,46 @@ public class UnitBooster : MonoBehaviour
                 Qx = 0;
             }
         }
-        Rt = transform.eulerAngles.z/90;
-        if((int)Rt == 0)
+        Rt = transform.localEulerAngles.z / 90;
+        if (B.boosters != null)
         {
-            B.GetComponent<AttackUnit>().boosters[0].Add(transform);
-        }
-        else if ((int)Rt == 1)
-        {
-            B.GetComponent<AttackUnit>().boosters[4].Add(transform);
-        }
-        else if ((int)Rt == 2)
-        {
-            B.GetComponent<AttackUnit>().boosters[1].Add(transform);
-        }
-        else if ((int)Rt == 3)
-        {
-            B.GetComponent<AttackUnit>().boosters[5].Add(transform);
-        }
-        if(Qx != 0)
-        {
-            List<int> temp = Q.GetRange(0, 4-Qx);
-            Q.RemoveRange(0, 4-Qx);
-            Q.AddRange(temp);
-            if(Q[(int)Rt] == 0)
+            if ((int)Rt == 0)
             {
-                B.GetComponent<AttackUnit>().boosters[2].Add(transform);
+                B.boosters[0].Add(transform);
             }
-            else
+            else if ((int)Rt == 1)
             {
-                B.GetComponent<AttackUnit>().boosters[3].Add(transform);
+                B.boosters[4].Add(transform);
+            }
+            else if ((int)Rt == 2)
+            {
+                B.boosters[1].Add(transform);
+            }
+            else if ((int)Rt == 3)
+            {
+                B.boosters[5].Add(transform);
+            }
+            if (Qx != 0)
+            {
+                List<int> temp = Q.GetRange(0, 4 - Qx);
+                Q.RemoveRange(0, 4 - Qx);
+                Q.AddRange(temp);
+                if (Q[(int)Rt] == 0)
+                {
+                    B.boosters[2].Add(transform);
+                }
+                else
+                {
+                    B.boosters[3].Add(transform);
+                }
+                if ((int)Rt == 0 && Qx == 3)
+                {
+                    B.boosters[4].Add(transform);
+                }
+                if ((int)Rt == 0 && Qx == 4)
+                {
+                    B.boosters[5].Add(transform);
+                }
             }
         }
     }
